+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- **Plugins**:
+    - `gcp://`: Google Cloud Secret Manager source implementation (available in `plugin/source/gcp`).
 - **Built-in Modifiers**:
     - `?b64`: Encode secret value to a base64 string.
 - **Documentation**: